Ng-include & Child Controller & Scope problem is fixed
Problem arised with the use of newer version of ui-bootstrap: After
login the ideas page will be called. On the top of the ideas page we
have an ng-include statement for the navBar header. Normally this
header’s own controller should be called each time when its html is
included. It is apparently not so. So if we need sth refreshed from
this controller, we have to write it as a callback function, not as a
scope variable.